Fixed type of ExtentStripeLength
Pegasus returned cryptic CIMError: (1, '61') if it was wrong.

+++ providers/ExtentProvider.py
@@ -176,7 +176,7 @@ class ExtentProvider(DeviceProvider):
         model['NoSinglePointOfFailure'] = redundancy.no_single_point_of_failure
         model['DataRedundancy'] = pywbem.Uint16(redundancy.data_dedundancy)
         model['PackageRedundancy'] = pywbem.Uint16(redundancy.package_redundancy)
-        model['ExtentStripeLength'] = pywbem.Uint16(redundancy.stripe_length)
+        model['ExtentStripeLength'] = pywbem.Uint64(redundancy.stripe_length)
         model['IsComposite'] = (len(device.parents) > 1)
 
         # TODO: add DeltaReservation (mandatory in SMI-S)
